refactor(bunkershot3d): split the field schema by concern, not by size
fields/schema.py hit 1218 lines against the repo 1200-line budget. The
split is by concern rather than a suppression, because the concerns were
already distinct and the size was the symptom.

- schema.py keeps what a field CONTAINS: the layout and quantity
  vocabulary, the sampling geometry, the frame and series containers, and
  the digest. 547 lines.
- standing.py takes what a field CLAIMS: which tier produced it, how
  valid that answer is, where it says there is sand at all, and what was
  dropped to store it. 555 lines.
- units.py holds the four SI unit strings, so the container and the
  standing can each name a unit without importing the other and no view
  has a second copy of "kg/m^3" waiting to drift.

Contents and claims change for different reasons. Adding a quantity is a
change to the container; a new tier or a new validity rule is a change to
the claims. Neither edit now has to open the other file.

FIELD_SCHEMA_VERSION moves to standing.py, because a format version is a
claim a file makes about itself and because leaving it beside the arrays
would have made the container import the standing and the standing import
the container.

No behaviour change: fields/__init__.py exports exactly the same names.

Part of #8699, #8710, #8711.
