+++ src/bunkershot3d/fields/store.py
@@ -44,15 +44,17 @@ from ..provenance.manifest import RunManifest, Validity
 from ..provenance.rng import SeedRecord, root_seed_sequence, seed_record
 from ..solvers.envelope import EnvelopeStatus
 from .schema import (
-    FIELD_SCHEMA_VERSION,
-    FieldIntegrityError,
     FieldLayout,
-    FieldProvenance,
     GridGeometry,
-    OccupancyRule,
-    RetentionRecord,
     SandFieldSeries,
     series_digest,
+)
+from .standing import (
+    FIELD_SCHEMA_VERSION,
+    FieldIntegrityError,
+    FieldProvenance,
+    OccupancyRule,
+    RetentionRecord,
 )
 
 __all__ = [
